Sound handling of originHost
I removed the getLocalHostName() method call in the constructur of GelfAppender. I did this because on my local machine this called resulted in an exeception due to the local hostname could not be resolved. As such a scenario might also be happening in a production enviroment I propose that the originHost should be set via a static method. For this I changed the classes GelfAppender and GelfConsoleAppender.