GELF timestamp should include milliseconds
The internal timestamp in Graylog 0.9.6 appears to be
seconds[.thousandths of a second] format, in contrast to the GELF spec
@ [1] which states that the timestamp field  is in microseconds (Graylog
is only off by 6 orders of magnitude :)).

Changed setTimestamp to setJavaTimestamp to be consistent with the
constructor, which takes a java timestamp. The internal GELF timestamp
cannot be set from outside the class.

The change here works with Graylog 0.9.6 and correctly interprets the date
value with milliseconds precision.

[1] https://github.com/Graylog2/graylog2-docs/wiki/GELF