Configure patterns in DefaultGelfMessageFactory
Sometimes it is desirable to use the default marshaller, but just
change the message patterns. For example, one might want to include the
log event message in full message, and trim the short message to 50
characters at the most. Prior to this change it required implementing a
custom marshaller or using reflection to lookup the pattern layouts and
reconfigure them.

With this change, setting custom patterns is a first-class citizen.